requirements.txt
cassandra-driver==3.6.0
Django==1.10
django-cassandra-engine==0.11.0
futures==3.0.5
six==1.10.0
wheel==0.24.0
